feat: add explorer logic start

+++ src/sim/base/components.rs
@@ -2,7 +2,7 @@ use bevy::prelude::*;
 
 #[derive(Component)]
 pub struct Base {
-    pub pos: (i32, i32),
+    pub pos: (f32, f32),
     pub iron: u32,
     pub nb_explorer_max: usize,
 }
